requirements.txt
celery==3.1.24
Django==1.10.2
django-filter==0.15.3
djangorestframework==3.4.7
kombu==3.0.37
Pillow==3.4.2
requests==2.11.1
